+++ project/apps/user/src/app/authentication/authentication.http
@@ -5,7 +5,7 @@ Content-Type: application/json
 {
   "firstname": "John",
   "lastname": "Doe",
-  "email": "john5453@doe.com",
+  "email": "john95854@doe.com",
   "cityId": 2,
   "roleId": 1,
   "password": "123456",
@@ -19,36 +19,19 @@ POST http://localhost:3000/api/v1/auth/login HTTP/1.1
 Content-Type: application/json
 
 {
-  "email": "john5453@doe.com",
+  "email": "john95854@doe.com",
   "password": "123456"
 }
 
 
 
 ### Изменение пользовательского пароля
-PATCH  http://localhost:3000/api/v1/auth/password/963d1b4e-2979-403e-a107-4bb40f7d1cb3 HTTP/1.1
+PATCH  http://localhost:3000/api/v1/auth/password/2a51e38a-2614-4907-ae7a-c706c00b919b HTTP/1.1
 Content-Type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIyZGJkN2UwYy03Zjg4LTQ3YmQtYjE4My1lNjE5NGFlNWM0MTkiLCJlbWFpbCI6ImpvaG45M0Bkb2UuY29tIiwicm9sZUlkIjoxLCJsYXN0bmFtZSI6IkRvZSIsImZpcnN0bmFtZSI6IkpvaG4iLCJpYXQiOjE3MDQ3MTQ2MTcsImV4cCI6MTcwNDcxODIxN30.jfPH-QXRLxNborqWBgWUj0xy4rTFeaF1vxGvpihn1cw
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIyYTUxZTM4YS0yNjE0LTQ5MDctYWU3YS1jNzA2YzAwYjkxOWIiLCJlbWFpbCI6ImpvaG45NTg1NEBkb2UuY29tIiwicm9sZUlkIjoxLCJsYXN0bmFtZSI6IkRvZSIsImZpcnN0bmFtZSI6IkpvaG4iLCJpYXQiOjE3MDUwNDkwMTcsImV4cCI6MTcwNTA0OTMxN30.pk80-RK_mKG18TuJHan4WP2krTk2pGIEnX0zTt70sWE
 
 {
   "oldPassword": "123456",
   "newPassword": "567890"
 }
 
-
-
-### Выход из авторизованной зоны
-POST  http://localhost:3000/api/v1/auth/logout HTTP/1.1
-Content-Type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIyZGJkN2UwYy03Zjg4LTQ3YmQtYjE4My1lNjE5NGFlNWM0MTkiLCJlbWFpbCI6ImpvaG45M0Bkb2UuY29tIiwicm9sZUlkIjoxLCJsYXN0bmFtZSI6IkRvZSIsImZpcnN0bmFtZSI6IkpvaG4iLCJpYXQiOjE3MDQ3MTQ2MTcsImV4cCI6MTcwNDcxODIxN30.jfPH-QXRLxNborqWBgWUj0xy4rTFeaF1vxGvpihn1cw
-
-
-
-### Обновление токенов
-POST  http://localhost:3000/api/v1/auth/tokens HTTP/1.1
-Content-Type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIyZGJkN2UwYy03Zjg4LTQ3YmQtYjE4My1lNjE5NGFlNWM0MTkiLCJlbWFpbCI6ImpvaG45M0Bkb2UuY29tIiwicm9sZUlkIjoxLCJsYXN0bmFtZSI6IkRvZSIsImZpcnN0bmFtZSI6IkpvaG4iLCJpYXQiOjE3MDQ3MTQ2MTcsImV4cCI6MTcwNDcxODIxN30.jfPH-QXRLxNborqWBgWUj0xy4rTFeaF1vxGvpihn1cw
-
-{
-  "refreshToken": "123fsd4fk4jf43sd4kr43kd4ff4ds"
-}
