--- project/apps/user/src/app/authentication/dto/logout-user.dto.ts
@@ -1,14 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-import { IsString, IsJWT, IsNotEmpty } from 'class-validator';
-
-export class LogoutUserDto {
-  @ApiProperty({
-    description: 'Access token',
-    example:
-      'eyJhbGciOiJIUzUxMiIsI.hgJhbGciOiTOOzUxMiIsI.ytJhbGciOiJgreUzUxMiIsI',
-  })
-  @IsNotEmpty()
-  @IsString()
-  @IsJWT()
-  public accessToken: string;
-}
