+++ project/apps/user/src/app/users/users.http
@@ -1,16 +1,17 @@
 ### Получение детальной информации о пользователе
-GET http://localhost:3000/api/v1/users/963d1b4e-2979-403e-a107-4bb40f7d1cb3 HTTP/1.1
+GET http://localhost:3000/api/v1/users/6d939de8-8801-461c-a567-de5b060a158c HTTP/1.1
 Content-Type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIyZGJkN2UwYy03Zjg4LTQ3YmQtYjE4My1lNjE5NGFlNWM0MTkiLCJlbWFpbCI6ImpvaG45M0Bkb2UuY29tIiwicm9sZUlkIjoxLCJsYXN0bmFtZSI6IkRvZSIsImZpcnN0bmFtZSI6IkpvaG4iLCJpYXQiOjE3MDQ3MTQ2MTcsImV4cCI6MTcwNDcxODIxN30.jfPH-QXRLxNborqWBgWUj0xy4rTFeaF1vxGvpihn1cw
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI0MjBkOWUwMy1mMjEyLTRjNzEtYjRlZS1mYzcxZjY1MTM2OGMiLCJlbWFpbCI6ImpvaG41NDUzMkBkb2UuY29tIiwicm9sZUlkIjoxLCJsYXN0bmFtZSI6IkRvZSIsImZpcnN0bmFtZSI6IkpvaG4iLCJpYXQiOjE3MDQ5MTgwNDEsImV4cCI6MTcwNDkxODM0MX0.WssP5E_Tw898U8HT-rE97Dzs-D_anl-PbDaPIBwDYu0
 
 
 
 ### Изменение информации профиля
-PATCH  http://localhost:3000/api/v1/users/963d1b4e-2979-403e-a107-4bb40f7d1cb3/profile HTTP/1.1
+PATCH  http://localhost:3000/api/v1/users/6d939de8-8801-461c-a567-de5b060a158c/profile HTTP/1.1
 Content-Type: application/json
-Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIyZGJkN2UwYy03Zjg4LTQ3YmQtYjE4My1lNjE5NGFlNWM0MTkiLCJlbWFpbCI6ImpvaG45M0Bkb2UuY29tIiwicm9sZUlkIjoxLCJsYXN0bmFtZSI6IkRvZSIsImZpcnN0bmFtZSI6IkpvaG4iLCJpYXQiOjE3MDQ3MTQ2MTcsImV4cCI6MTcwNDcxODIxN30.jfPH-QXRLxNborqWBgWUj0xy4rTFeaF1vxGvpihn1cw
+Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI0MjBkOWUwMy1mMjEyLTRjNzEtYjRlZS1mYzcxZjY1MTM2OGMiLCJlbWFpbCI6ImpvaG41NDUzMkBkb2UuY29tIiwicm9sZUlkIjoxLCJsYXN0bmFtZSI6IkRvZSIsImZpcnN0bmFtZSI6IkpvaG4iLCJpYXQiOjE3MDQ5MTgwNDEsImV4cCI6MTcwNDkxODM0MX0.WssP5E_Tw898U8HT-rE97Dzs-D_anl-PbDaPIBwDYu0
 
 {
   "firstname": "Jack",
-  "lastname": "Russel"
+  "lastname": "Russel",
+  "info": "Some text"
 }
