+++ project/apps/user/src/app/users/users.validation.ts
@@ -1,5 +1,5 @@
 export enum UserRule {
-  NameMinLength = 2,
+  NameMinLength = 3,
   NameMaxLength = 50,
   DescriptionMaxLength = 300,
   MinAge = 18,
